gateway/src/main/webapp/app/entities/xaracter/route/xaracter-routing-resolve.service.spec.ts
import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, ActivatedRoute, Router, convertToParamMap } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of } from 'rxjs';

import { IXaracter, Xaracter } from '../xaracter.model';
import { XaracterService } from '../service/xaracter.service';

import { XaracterRoutingResolveService } from './xaracter-routing-resolve.service';

describe('Xaracter routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let routingResolveService: XaracterRoutingResolveService;
  let service: XaracterService;
  let resultXaracter: IXaracter | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
      providers: [
        {
          provide: ActivatedRoute,
          useValue: {
            snapshot: {
              paramMap: convertToParamMap({}),
            },
          },
        },
      ],
    });
    mockRouter = TestBed.inject(Router);
    jest.spyOn(mockRouter, 'navigate').mockImplementation(() => Promise.resolve(true));
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRoute).snapshot;
    routingResolveService = TestBed.inject(XaracterRoutingResolveService);
    service = TestBed.inject(XaracterService);
    resultXaracter = undefined;
  });

  describe('resolve', () => {
    it('should return IXaracter returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultXaracter = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultXaracter).toEqual({ id: 123 });
    });

    it('should return new IXaracter if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultXaracter = result;
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultXaracter).toEqual(new Xaracter());
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse({ body: null as unknown as Xaracter })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultXaracter = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultXaracter).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
